fix: 위치 main으로 수정
