+++ web/app/onboarding/terms/page.tsx
@@ -1,12 +1,10 @@
 'use client';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import ProgressBar from '@/components/common/ProgressBar';
 import Button from '@/components/common/Button';
 import { useRouter } from 'next/navigation';
 import OnBoardingWrapper from '@/components/onBoarding/OnBoardingWrapper';
 import TermsAgreements from '@/components/onBoarding/TermsAgreements';
-import { postMessageToApp } from '@/utils/apis/postMessageToApp';
-import { MODULE } from '@/utils/apis/api';
 
 function Page() {
   const [isButtonDisabled, setIsButtonDisabled] = useState(true);
@@ -15,26 +13,6 @@ function Page() {
   const handleAgreementChange = (isValid: boolean) => {
     setIsButtonDisabled(!isValid);
   };
-  useEffect(() => {
-    const cookie = document.cookie;
-    const cookieArray = cookie.split(';');
-    const accessToken = cookieArray.find(cookie =>
-      cookie.startsWith('accessToken=')
-    );
-    const refreshToken = cookieArray.find(cookie =>
-      cookie.startsWith('refreshToken=')
-    );
-    const refreshTokenValue = refreshToken?.split('=')[1];
-    const accessTokenValue = accessToken?.split('=')[1];
-
-    postMessageToApp(
-      MODULE.AUTH,
-      JSON.stringify({
-        accessToken: accessTokenValue,
-        refreshToken: refreshTokenValue
-      })
-    );
-  }, []);
 
   return (
     <div className="flex flex-grow flex-col">
